requirements.txt
# Core dependencies
openai>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Discord integration
py-cord>=2.5.0
discord-py>=2.4.0

# GitHub integration
PyGithub>=2.3.0
GitPython>=3.1.40

# Docker integration
docker>=6.1.0

# Configuration management
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Async utilities
aiohttp>=3.9.0
aiofiles>=23.2.0

# Logging and monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0